separate out the comment classes into their own files
